% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvivalAnalysis.R
\name{ILDREA}
\alias{ILDREA}
\title{Survival and Recurrent Events Analysis for Intensive Longitudinal EMA Data}
\usage{
ILDREA(directory)
}
\arguments{
\item{directory}{The directory where the EMA dataframe is available.}
}
\description{
This function produces analysis of effects on time to next use for intensive longitudinal EMA data.
}
\examples{
ILDREA()
}
\keyword{Data}
\keyword{EMA}
\keyword{Intensive}
\keyword{Longitudinal}
\keyword{Survival}
