% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvivalDataFormat.R
\name{ILDREAFormat}
\alias{ILDREAFormat}
\title{Survival and Recurrent Events Analysis Data Formatting for Intensive Longitudinal EMA Data}
\usage{
ILDREAFormat(directory)
}
\arguments{
\item{directory}{The directory where the EMA dataframe is available.}
}
\description{
This function prepares the data for analysis of effects on time to next use for intensive longitudinal EMA data.
}
\examples{
ILDREAFormat()
}
\keyword{Data}
\keyword{EMA}
\keyword{Intensive}
\keyword{Longitudinal}
\keyword{Survival}
