% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmoothedPlots.R
\name{SmoothPlots}
\alias{SmoothPlots}
\title{Smoothed Plots for Intensive Longitudinal Functional Data}
\usage{
SmoothPlots(directory)
}
\arguments{
\item{directory}{The directory where the functional dataframe is available.}
}
\description{
This function produces loess smoothed plots for one school data of activity data.
}
\examples{
SmoothPlots()
}
\keyword{Data}
\keyword{Functional}
\keyword{Intensive}
\keyword{Longitudinal}
\keyword{Smooth}
