% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSinceUse.R
\name{TimeSinceUse}
\alias{TimeSinceUse}
\title{Time Since Last Use Analysis for Intensive Longitudinal EMA Data}
\usage{
TimeSinceUse(directory)
}
\arguments{
\item{directory}{The directory where the EMA dataframe is available.}
}
\description{
This function produces analysis of time since last use for intensive longitudinal EMA data.
}
\examples{
TimeSinceUse()
}
\keyword{Data}
\keyword{EMA}
\keyword{Intensive}
\keyword{Longitudinal}
\keyword{Time}
